update docker file. restructure folder structure. minimal renamings and defaults
